Database connection and di was completed

+++ app/src/main/java/com/marazanil/personsapplication/ui/viewmodel/PersonDetailFragmentViewModel.kt
@@ -8,7 +8,7 @@ import javax.inject.Inject
 @HiltViewModel
 class PersonDetailFragmentViewModel @Inject constructor(var personRepo : PersonDaoRepository): ViewModel(){
 
-    fun update(personId:Long, personDetailName:String, personPhoneNumberDetail : String){
+    fun update(personId:Int, personDetailName:String, personPhoneNumberDetail : String){
         personRepo.personUpdate(personId,personDetailName,personPhoneNumberDetail)
     }
